# artifacts/architect/openapi.yaml
openapi: 3.0.3
info:
  title: DriveMind API - Production Repair Analysis
  description: |
    **PRODUCTION DEPLOYED** Google Drive management platform with CRITICAL FUNCTIONALITY ISSUES.
    
    ## PRODUCTION STATUS (v3.0.0-REPAIR)
    - **DEPLOYMENT**: ✅ Live at https://studio--drivemind-q69b7.us-central1.hosted.app
    - **AUTHENTICATION**: ✅ Google OAuth working
    - **CORE FEATURES**: ❌ BROKEN (Background scans, real inventory, file operations)
    - **USER EXPERIENCE**: ❌ NON-FUNCTIONAL (Users see stubbed data only)
    
    ## CRITICAL ISSUES IDENTIFIED
    - **Background Scans**: Start but never complete, jobs stuck in running state
    - **File Inventory**: Using mock data instead of real Google Drive files
    - **File Operations**: All APIs disabled (in api-disabled folder)
    - **Real-time Updates**: UI shows indefinite loading states
    - **Token Persistence**: Authentication may not persist reliably
    
    ## IMMEDIATE REPAIR REQUIRED
    Emergency fixes needed within 48-72 hours to restore basic functionality.
    Core user journey (authenticate → scan → view files → manage) currently broken.
    
    ## Rate Limits (CURRENT)
    - OAuth endpoints: 50 requests/minute per IP
    - Workflow endpoints: 30 requests/minute per user  
    - AI endpoints: 20 requests/minute per user
    - Health/metrics: 100 requests/minute per IP
    
  version: 3.0.0-REPAIR
  contact:
    name: DriveMind Support
    email: support@drivemind.ai
  license:
    name: MIT
...
servers:
...
security:
  - BearerAuth: []
  - CookieAuth: []
paths:
...
  /workflows/background-scan:
    post:
      summary: Start background drive scan
      description: |
        Initiates an asynchronous drive scan using Cloud Functions.
        Returns immediately with scan ID for status tracking.
      operationId: backgroundScan
      tags:
        - Workflows
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - userId
              properties:
                userId:
                  type: string
                  description: User identifier
                maxDepth:
                  type: integer
                  minimum: 1
                  maximum: 50
                  default: 20
                includeTrashed:
                  type: boolean
                  default: false
                scanSharedDrives:
                  type: boolean
                  default: false
      responses:
        '202':
          description: Background scan initiated
          content:
            application/json:
              schema:
                type: object
                properties:
                  scanId:
                    type: string
                    description: Unique scan identifier
                  status:
                    type: string
                    enum: [queued, running]
                  estimatedCompletion:
                    type: string
                    format: date-time
                    description: Estimated completion time
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase ID token
    CookieAuth:
      type: apiKey
      in: cookie
      name: google_access_token
      description: HTTP-only OAuth access token cookie
  schemas:
    Error:
      type: object
      required:
        - error
        - message
        - timestamp
      properties:
        error:
          type: string
          description: Error code
          example: "invalid_request"
        message:
          type: string
          description: Human-readable error message
          example: "Required field 'userId' is missing"
        details:
          type: string
          description: Additional error details
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
          description: Unique request identifier for debugging
...
  responses:
    BadRequest:
      description: Invalid request data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            validation_error:
              summary: Validation failure
              value:
                error: "invalid_request"
                message: "Request validation failed"
                details: "Required field 'userId' is missing"
                timestamp: "2024-12-13T10:30:00Z"
                requestId: "req_123abc"
    Unauthorized:
      description: Authentication required or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            missing_auth:
              summary: Missing authentication
              value:
                error: "unauthorized"
                message: "No authorization token provided"
                timestamp: "2024-12-13T10:30:00Z"
            invalid_token:
              summary: Invalid token
              value:
                error: "invalid_token"
                message: "Authorization token is expired or invalid"
                timestamp: "2024-12-13T10:30:00Z"
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            service_error:
              summary: Service unavailable
              value:
                error: "internal_server_error"
                message: "An unexpected error occurred"
                timestamp: "2024-12-13T10:30:00Z"
                requestId: "req_456def"
            dependency_failure:
              summary: External service failure
              value:
                error: "service_unavailable"
                message: "Google Drive API is temporarily unavailable"
                details: "Upstream service returned 503"
                timestamp: "2024-12-13T10:30:00Z"
...